+++ src/legend/interactiveLegend.ts
@@ -24,6 +24,7 @@
 *  THE SOFTWARE.
 */
 
+/* eslint-disable powerbi-visuals/no-implied-inner-html */
 import { select, Selection } from "d3-selection";
 
 import powerbi from "powerbi-visuals-api";
@@ -58,7 +59,7 @@ export class InteractiveLegend implements ILegend {
     }
 
     public drawLegend(legendData: LegendData) {
-        let data = legendData.dataPoints;
+        const data = legendData.dataPoints;
         if (data.length < 1) {
             return;
         }
@@ -97,6 +98,7 @@ export class InteractiveLegend implements ILegend {
     /**
      * Not supported
      */
+    /* eslint-disable-next-line @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function */
     public changeOrientation(orientation: LegendPosition) { }
 
     public getOrientation(): LegendPosition {
@@ -107,11 +109,11 @@ export class InteractiveLegend implements ILegend {
      * Draw the legend title
      */
     private drawTitle(data: LegendDataPoint[]): void {
-        let titleDiv: Selection<any, any, any, any> = this.legendContainerDiv.selectAll(`div.${InteractiveLegend.LegendTitleClass}`),
+        const titleDiv: Selection<any, any, any, any> = this.legendContainerDiv.selectAll(`div.${InteractiveLegend.LegendTitleClass}`),
             item: Selection<any, any, any, any> = titleDiv.data([data[0]]);
 
         // Enter
-        let itemEnter: Selection<any, any, any, any> = item.enter(),
+        const itemEnter: Selection<any, any, any, any> = item.enter(),
             titleDivEnter: Selection<any, any, any, any> = itemEnter
                 .append("div")
                 .attr("class", InteractiveLegend.LegendTitleClass);
@@ -131,9 +133,9 @@ export class InteractiveLegend implements ILegend {
             .style(InteractiveLegend.legendColorCss, (d: LegendDataPoint) => d.color);
 
         item
-        .merge(itemEnter)
-        .select("span:last-child")
-        .text((d: LegendDataPoint) => d.category);
+            .merge(itemEnter)
+            .select("span:last-child")
+            .text((d: LegendDataPoint) => d.category);
     }
 
     /**
@@ -143,24 +145,24 @@ export class InteractiveLegend implements ILegend {
         // Add Mesaures - the items of the category in the legend
         this.ensureLegendTableCreated();
 
-        let dataPointsMatrix: LegendDataPoint[][] = [data];
-        let legendItemsContainer: Selection<any, any, any, any> = this.legendContainerDiv
+        const dataPointsMatrix: LegendDataPoint[][] = [data];
+        const legendItemsContainer: Selection<any, any, any, any> = this.legendContainerDiv
             .select("tbody")
             .selectAll("tr")
             .data(dataPointsMatrix);
 
         // Enter
-        let legendItemsEnter: Selection<any, any, any, any> = legendItemsContainer.enter(),
+        const legendItemsEnter: Selection<any, any, any, any> = legendItemsContainer.enter(),
             rowEnter: Selection<any, any, any, any> = legendItemsEnter.append("tr");
 
-        let cellEnter: Selection<any, any, any, any> = rowEnter
+        const cellEnter: Selection<any, any, any, any> = rowEnter
             .selectAll("td")
             .data((d: LegendDataPoint[]) => d, (d: LegendDataPoint) => d.label)
             .enter()
             .append("td")
             .attr("class", InteractiveLegend.LegendItem);
 
-        let cellSpanEnter: Selection<any, any, any, any> = cellEnter.append("span");
+        const cellSpanEnter: Selection<any, any, any, any> = cellEnter.append("span");
 
         cellSpanEnter.filter((d: LegendDataPoint) => !d.iconOnlyOnLabel)
             .append("span")
@@ -183,7 +185,7 @@ export class InteractiveLegend implements ILegend {
             .attr("class", InteractiveLegend.legendItemMeasureClass);
 
         // Update
-        let legendCells: Selection<any, any, any, any> = legendItemsContainer
+        const legendCells: Selection<any, any, any, any> = legendItemsContainer
             .merge(legendItemsEnter)
             .selectAll("td")
             .data((d: LegendDataPoint[]) => d, (d: LegendDataPoint) => d.label);
@@ -214,7 +216,7 @@ export class InteractiveLegend implements ILegend {
      */
     private ensureLegendTableCreated(): void {
         if (this.legendContainerDiv.select("div table").empty()) {
-            let legendTable: Selection<any, any, any, any> = this.legendContainerDiv
+            const legendTable: Selection<any, any, any, any> = this.legendContainerDiv
                 .append("div")
                 .append("table");
 
@@ -228,6 +230,8 @@ export class InteractiveLegend implements ILegend {
     /**
      * Set Horizontal Pan gesture for the legend
      */
+    
+    /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
     private setPanGestureOnLegend(legendTable: Selection<any, any, any, any>): void {
         throw "Not implemented";
         // let parentNode = <HTMLElement>this.legendContainerParent.node();